pyproject.toml
[tool.poetry]
name = "ai-test"
version = "1.0"
description = "'An application for demonstrating your skills in AI'"
authors = ["Rockstar"]

[tool.poetry.dependencies]
python = "^3.8"
openfabric-pysdk = "^0.3.0"
Pillow = ">=10.0.0,<11.0.0"
gradio = "3.50.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.source]]
name = "node2"
url = "https://repo.node2.openfabric.network/index"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
